introduced compile flag to encapsulate MPI
