+++ .gitignore
@@ -19,6 +19,8 @@ i10pc128/*
 i10pc129/*
 i10pc130/*
 i128/*
+i129/*
+mpii129/*
 mpii128/*
 debug2/*
 tuning/*
